+++ Client/Private/InventoryIcon.cpp
@@ -26,7 +26,7 @@ HRESULT CInventoryIcon::Initialize(void * pArg)
 	if (FAILED(__super::Initialize(pArg)))
 		return E_FAIL;
 
-	D3DXMatrixOrthoLH(&m_ProjMatrix, g_iWinSizeX, g_iWinSizeY, 0.f, 1.f);
+	D3DXMatrixOrthoLH(&m_ProjMatrix, (float)g_iWinSizeX, (float)g_iWinSizeY, 0.f, 1.f);
 	m_Click = (bool*)pArg;
 	m_fSizeX = 30.0f;
 	m_fSizeY = 30.0f;
@@ -48,7 +48,7 @@ void CInventoryIcon::Tick(_float fTimeDelta)
 	if (*m_Click == true)
 	{
 		RECT		rcRect;
-		SetRect(&rcRect, m_fX - m_fSizeX * 0.5f, m_fY - m_fSizeY * 0.5f, m_fX + m_fSizeX * 0.5f, m_fY + m_fSizeY * 0.5f);
+		SetRect(&rcRect, (int)(m_fX - m_fSizeX * 0.5f), (int)(m_fY - m_fSizeY * 0.5f), (int)(m_fX + m_fSizeX * 0.5f), (int)(m_fY + m_fSizeY * 0.5f));
 
 		POINT		ptMouse;
 		GetCursorPos(&ptMouse);
